rater/requirements.txt
aiohttp==3.6.3
asgiref==3.4.1
discord.py==1.7.1
Flask==2.0.1
fuzzywuzzy==0.18.0
heroku3==4.2.3
numpy==1.19.4
opencv-python-headless==4.4.0.46
psycopg2-binary==2.9.1
python-dotenv==0.15.0
python-Levenshtein==0.12.0
SQLAlchemy==1.3.22
Unidecode==1.1.1
validators==0.18.1
